limitando el tablero, agregando pregunta de si puede volver a comer y el comer de la reina
